*refactor sanity queries
