+++ sanity/data/footer.ts
@@ -1,0 +1,3 @@
+import { languageSupport, logoMark, navPage } from "./global";
+
+export const footerQuery = `*[_type=="footer"][0]{...,footerRoutes[]{...,routes[]{...,${navPage}}},policies[]{...,${navPage}},"langSupport":${languageSupport},"logo":${logoMark}}`;
