+++ sanity/queries/sitemap.ts
@@ -1,5 +1,5 @@
 import { SanityDocument } from "next-sanity";
-import { BLOGS_SITEMAP_QUERY, SITEMAP_QUERY } from "../data/queries";
+import { BLOGS_SITEMAP_QUERY, SITEMAP_QUERY } from "../data";
 import { sanityFetch } from "../lib/client";
 
 export const fetchPageSitemap = async () => {
